nip01 data layer refactor
