+++ packages/nip01/lib/nip01.dart
@@ -12,7 +12,8 @@ export 'src/data/models/key_pair.dart';
 export 'src/data/models/kind_0_metadata.dart';
 export 'src/data/models/relay_message.dart';
 
-export 'src/data/providers/relay_connection_provider.dart';
+export 'src/data/providers/relay_stream_provider.dart';
+export 'src/data/clients/relay_client.dart';
 
 export 'src/enums/event_kind.dart';
 export 'src/enums/client_message_type.dart';
