+++ packages/nip01/pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   crypto: ^3.0.3
   meta: ^1.15.0
   web_socket_channel: ^3.0.1
+  async: ^2.11.0
   # path: ^1.8.0
 
 dev_dependencies:
